Further loader refactor
- moved reader, dask, aws and rio to loader
  subpackage

- moved SomeReader into loader.types
