+++ odc/stac/loader/__init__.py
@@ -9,8 +9,13 @@ from ._builder import (
     fill_2d_slice,
     mk_dataset,
 )
+from .types import RasterBandMetadata, RasterLoadParams, RasterSource, SomeReader
 
 __all__ = (
+    "RasterBandMetadata",
+    "RasterLoadParams",
+    "RasterSource",
+    "SomeReader",
     "LoadChunkTask",
     "DaskGraphBuilder",
     "mk_dataset",
