+++ tests/test_dask.py
@@ -1,7 +1,7 @@
 from dask.base import tokenize
 from odc.geo.geobox import CRS, GeoBox, GeoboxTiles
 
-from odc.stac._dask import tokenize_stream
+from odc.stac.loader._dask import tokenize_stream
 
 
 def test_tokenize_odc_geo():
