+++ tests/test_utils_aws.py
@@ -8,7 +8,7 @@ from unittest import mock
 import pytest
 
 _ = pytest.importorskip("botocore")
-from odc.stac._aws import (
+from odc.stac.loader._aws import (
     _fetch_text,
     auto_find_region,
     ec2_current_region,
@@ -40,36 +40,42 @@ def test_ec2_current_region():
     ]
 
     for rv, expect in tests:
-        with mock.patch("odc.stac._aws._fetch_text", return_value=rv):
+        with mock.patch("odc.stac.loader._aws._fetch_text", return_value=rv):
             assert ec2_current_region() == expect
 
 
-@mock.patch("odc.stac._aws.botocore_default_region", return_value=None)
+@mock.patch("odc.stac.loader._aws.botocore_default_region", return_value=None)
 def test_auto_find_region(*mocks):
-    with mock.patch("odc.stac._aws._fetch_text", return_value=None):
+    with mock.patch("odc.stac.loader._aws._fetch_text", return_value=None):
         with pytest.raises(ValueError):
             auto_find_region()
 
-    with mock.patch("odc.stac._aws._fetch_text", return_value=_json(region="TT")):
+    with mock.patch(
+        "odc.stac.loader._aws._fetch_text", return_value=_json(region="TT")
+    ):
         assert auto_find_region() == "TT"
 
 
-@mock.patch("odc.stac._aws.botocore_default_region", return_value="tt-from-botocore")
+@mock.patch(
+    "odc.stac.loader._aws.botocore_default_region", return_value="tt-from-botocore"
+)
 def test_auto_find_region_2(*mocks):
     assert auto_find_region() == "tt-from-botocore"
 
 
 def test_fetch_text():
-    with mock.patch("odc.stac._aws.urlopen", return_value=mock_urlopen("", 505)):
+    with mock.patch("odc.stac.loader._aws.urlopen", return_value=mock_urlopen("", 505)):
         assert _fetch_text("http://localhost:8817") is None
 
-    with mock.patch("odc.stac._aws.urlopen", return_value=mock_urlopen("text", 200)):
+    with mock.patch(
+        "odc.stac.loader._aws.urlopen", return_value=mock_urlopen("text", 200)
+    ):
         assert _fetch_text("http://localhost:8817") == "text"
 
     def fake_urlopen(*args, **kw):
         raise IOError("Always broken")
 
-    with mock.patch("odc.stac._aws.urlopen", fake_urlopen):
+    with mock.patch("odc.stac.loader._aws.urlopen", fake_urlopen):
         assert _fetch_text("http://localhost:8817") is None
 
 
@@ -133,14 +139,16 @@ aws_secret_access_key = fake-fake-fake
     assert aws["region_name"] == "us-west-1"
     assert aws["aws_unsigned"] is True
 
-    with mock.patch("odc.stac._aws._fetch_text", return_value=_json(region="mordor")):
+    with mock.patch(
+        "odc.stac.loader._aws._fetch_text", return_value=_json(region="mordor")
+    ):
         aws, creds = get_aws_settings(profile="no_region", aws_unsigned=True)
 
         assert aws["region_name"] == "mordor"
         assert aws["aws_unsigned"] is True
 
 
-@mock.patch("odc.stac._aws.get_creds_with_retry", return_value=None)
+@mock.patch("odc.stac.loader._aws.get_creds_with_retry", return_value=None)
 def test_get_aws_settings_no_credentials(without_aws_env):
     # get_aws_settings should fail when credentials are not available
     with pytest.raises(ValueError, match="Couldn't get credentials"):
